fix: prevent project names from exploding the selection buttons
This change prevents long project names from blowing the form out of
proportion.

To do so, it:
sets `whitespace: no-wrap` on the button labels. Judging by the other
styles, this was the intention all along, but it didn't really come up
until now.

It also sets the label width for projects to 30ch or the longest
project name you have (whichever is shorter), so that you'll get
to see quite a bit of the project name before it gets cut off.