fix: correctly align filter button layouts
Fixes an issue where the filter buttons were both too far down and too
far to the right.

The issue was that the wrapper body imposed a pretty substantial bit
of padding. However, the filter buttons already came with their own
bit of padding. The result of this was alignment issues.

To fix it I have:
- set the default page body to have no padding.
- added a wrapper with padding around the event search results.

Ideally, I would have liked to have full control of the body
container (and the filter buttons) and solve this with margins.
However, with the building blocks we have, this was the simplest
solution I found. That said, I'd be happy to have suggestions for
better implementations.