fix: Update avatar usage in change requests
Turns out there was (at least) one other avatar instance we missed in
[ previous update ](https://github.com/Unleash/unleash/pull/7681):
change request avatars.

They don't use the existing UserAvatar component, so didn't show up in
the dependency graph. This change makes it so that those avatars also
use the user avatar component (and send all the info that we need)