chore: add flags `projectListImprovements` and `useProjectReadModel`
These are both related to the work on the project list improvements project.

The `projectListImprovements` flag will be used to enable disable the
new project list improvements.

The `useProjectReadModel` flag will be used to enable/disable the use
of the new project read model and is mostly a safety feature.