chore: enables github docker cache (#8187)
So, since we updated to build-push-action@v6 we're now getting a nice
summary of the build on the summary page of the action. It pointed out
that we're caching 0% of our docker build. Since we probably change our
own code more often than we change our dependencies, sharing the initial
yarn install step between builds from cache seems to be a good idea.